Добавлены коды ответов и разделение прав пользователей

+++ api/deleteBook.php
@@ -5,28 +5,34 @@ require_once 'core/core.php';
 use Bitrix\Books\BooksTable,
     Bitrix\Authors\AuthorsTable,
     Bitrix\Write\WriteTable;
-
-$data = json_decode(file_get_contents('php://input'), true);
-
-if(is_numeric($data["id"])){
-    BooksTable::delete($data["id"]);
-
-    $res = WriteTable::getList(array(
-        'select' => array('ID', 'BOOK_ID'),
-        'filter' => array('BOOK_ID' => $data["id"]),
-    ));
     
-    while($arr = $res->fetch()) {
-        WriteTable::delete($arr["ID"]);
+if($userLogin == ADMIN_LOGIN && $userPassword == ADMIN_PASSWORD){
+    $data = json_decode(file_get_contents('php://input'), true);
+
+    if(is_numeric($data["id"])){
+        BooksTable::delete($data["id"]);
+
+        $res = WriteTable::getList(array(
+            'select' => array('ID', 'BOOK_ID'),
+            'filter' => array('BOOK_ID' => $data["id"]),
+        ));
+        
+        while($arr = $res->fetch()) {
+            WriteTable::delete($arr["ID"]);
+        }
+
+        $arResult["message"] = $MESSAGES[200];
+        $CODE = 200;
+    } else {
+        $arResult["message"] = $MESSAGES[404];
+        $CODE = 404;
     }
-
-    $arResult["message"] = 'Книга успешно удалена';
-    $arResult["status"] = true;
+} elseif(!$userLogin && !$userPassword) {
+    $arResult["message"] = $MESSAGES[401];
+    $CODE = 401;
 } else {
-
-    $$arResult["message"] = 'Неверный формат id';
-    $arResult["status"] = false;
+    $arResult["message"] = $MESSAGES[403];
+    $CODE = 403;
 }
-
 
 ?>
